See description
AlliesManager created., it has a list for the allies an manages it.
EnemiesManager created, it has a list for the enemeis and the boss and manages them.
BattleEventsManager and BattleEventsHandler created, they have events and manage enter and exit combat,
Created Managers prefab, it has all (I think) the useful managers, be ware of copies in the scene.
Some mods to Stats, reincarnation and allySpawnerTest in ored to integrate new logic.